+++ Changelog.md
@@ -2,6 +2,16 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release. 
 
+## 0.3.0
+
+### Added
+
+- Exception listener
+
+### Changed
+
+- We do not care about kernel:debug when evaluating `happyr_api.wsse.debug`.
+
 ## 0.2.0
 
 - Removed controller and created `ResponseFactory`